+++ assets/cke5profiles.js
@@ -26,7 +26,7 @@ const cke5profiles = {
                 "value": "50",
                 "icon": "medium"
             }, {"name": "imageResize:75", "value": "75", "icon": "large"}],
-            "toolbar": ["imageTextAlternative", "|", "imageStyle:full", "imageStyle:alignLeft", "imageStyle:alignRight", "imageStyle:alignCenter", "|", "linkImage", "|", "imageResize:original", "imageResize:25", "imageResize:50", "imageResize:75"],
+            "toolbar": ["imageTextAlternative", "|", "imageStyle:block", "imageStyle:alignLeft", "imageStyle:alignRight", "imageStyle:alignCenter", "|", "linkImage", "|", "imageResize:original", "imageResize:25", "imageResize:50", "imageResize:75"],
             "styles": ["full", "alignLeft", "alignRight", "alignCenter"]
         },
         "table": {"contentToolbar": ["tableColumn", "tableRow", "mergeTableCells"]},